Refactor: 커스텀훅으로 관심사 분리
