+++ src/components/MyExperiences/skeleton-ui.tsx
@@ -1,0 +1,11 @@
+const SkeletonUi = () => {
+  return (
+    <div className="flex flex-col gap-4">
+      <div className="w-full h-[177px] tablet:h-[139px] bg-gray-50 rounded-xl"></div>
+      <div className="w-full h-[177px] tablet:h-[139px] bg-gray-50 rounded-xl"></div>
+      <div className="w-full h-[177px] tablet:h-[139px] bg-gray-50 rounded-xl"></div>
+    </div>
+  );
+};
+
+export default SkeletonUi;
