+++ src/components/ReservationList/Card/reservation-card.tsx
@@ -1,11 +1,8 @@
 import type { components } from "../../../types/api-types";
 import { dateCalc } from "../../../utils/date-time";
-import { filterList } from "../filter-data";
-import { useState } from "react";
-import { ConfirmModal } from "../../Modal";
 import ReviewModal from "../../Review/review";
-import { useCancelMyReservation } from "../../../hooks/mutations";
-import { useQueryClient } from "@tanstack/react-query";
+import { ConfirmModal } from "../../Modal";
+import { useReservationCard } from "./useReservationCard";
 
 const ReservationCard = (
   props: components["schemas"]["ReservationWithActivityResponseDto"]
@@ -21,30 +18,15 @@ const ReservationCard = (
     headCount,
     reviewSubmitted,
   } = props;
-  const accessToken = localStorage.getItem("accessToken");
-  const cancelMutation = useCancelMyReservation(accessToken!);
-  const queryClient = useQueryClient();
-  const filterStatus = filterList.find((data) => data.status === status);
-  const [isOpen, setIsOpen] = useState({
-    cancel: false,
-    review: false,
-  });
 
-  const handleModalOpen = (e: React.MouseEvent<HTMLButtonElement>) => {
-    const modal = e.currentTarget.name as keyof typeof isOpen;
-
-    setIsOpen((prevOpen) => ({
-      ...prevOpen,
-      [modal]: true,
-    }));
-  };
-
-  const handleModalClose = (modal: keyof typeof isOpen) => {
-    setIsOpen((prevOpen) => ({
-      ...prevOpen,
-      [modal]: false,
-    }));
-  };
+  const [
+    cancelMutation,
+    queryClient,
+    filterStatus,
+    isOpen,
+    handleModalOpen,
+    handleModalClose,
+  ] = useReservationCard(status);
 
   return (
     <div className="flex flex-col items-start gap-3">
@@ -78,30 +60,23 @@ const ReservationCard = (
               </div>
               <div className="flex gap-2 tablet:hidden">
                 {filterStatus?.status === "pending" && (
-                  <>
-                    <button
-                      type="button"
-                      onClick={() => console.log("예약 변경")}
-                      className="px-2.5 py-1.5 rounded-[8px] border border-gray-50 bg-white ty-14_M text-gray-600 cursor-pointer"
-                    >
-                      예약 변경
-                    </button>
-                    <button
-                      type="button"
-                      name="cancel"
-                      onClick={handleModalOpen}
-                      className="px-2.5 py-1.5 rounded-[8px] border border-gray-50 bg-gray-50 ty-14_M text-gray-600 cursor-pointer"
-                    >
-                      예약 취소
-                    </button>
-                  </>
+                  <button
+                    type="button"
+                    name="cancel"
+                    onClick={handleModalOpen}
+                    className="px-2.5 py-1.5 rounded-[8px] border border-gray-50 bg-gray-50 ty-14_M text-gray-600 cursor-pointer
+                      transition transform duration-200 hover:scale-105 hover:shadow-md"
+                  >
+                    예약 취소
+                  </button>
                 )}
                 {filterStatus?.status === "completed" && !reviewSubmitted && (
                   <button
                     type="button"
                     name="review"
                     onClick={handleModalOpen}
-                    className="px-2.5 py-1.5 rounded-[8px] bg-primary-500 ty-14_M text-white cursor-pointer"
+                    className="px-2.5 py-1.5 rounded-[8px] bg-primary-500 ty-14_M text-white cursor-pointer
+                    transition transform duration-200 hover:scale-105 hover:shadow-md"
                   >
                     후기 작성
                   </button>
@@ -119,30 +94,23 @@ const ReservationCard = (
       </div>
       <div className="hidden w-full tablet:flex tablet:gap-3">
         {filterStatus?.status === "pending" && (
-          <>
-            <button
-              type="button"
-              onClick={() => console.log("예약변경")}
-              className="grow p-2.5 rounded-[8px] border border-gray-50 bg-white ty-14_M text-gray-600 text-center cursor-pointer"
-            >
-              예약 변경
-            </button>
-            <button
-              type="button"
-              name="cancel"
-              onClick={handleModalOpen}
-              className="grow p-2.5 rounded-[8px] border border-gray-50 bg-gray-50 ty-14_M text-gray-600 cursor-pointer"
-            >
-              예약 취소
-            </button>
-          </>
+          <button
+            type="button"
+            name="cancel"
+            onClick={handleModalOpen}
+            className="grow p-2.5 rounded-[8px] border border-gray-50 bg-gray-50 ty-14_M text-gray-600 cursor-pointer
+              transition transform duration-200 hover:scale-105 hover:shadow-md"
+          >
+            예약 취소
+          </button>
         )}
         {filterStatus?.status === "completed" && !reviewSubmitted && (
           <button
             type="button"
             name="review"
             onClick={handleModalOpen}
-            className="grow p-2.5 rounded-[8px] bg-primary-500 ty-14_M text-white cursor-pointer"
+            className="grow p-2.5 rounded-[8px] bg-primary-500 ty-14_M text-white cursor-pointer
+            transition transform duration-200 hover:scale-105 hover:shadow-md"
           >
             후기 작성
           </button>
