+++ src/components/ReservationStatus/Dropdown/useScheduleDropdown.ts
@@ -1,0 +1,58 @@
+import { useEffect, useRef, useState } from "react";
+
+type ScheduleDropdownType = (
+  handleSchedule: (id: number) => void
+) => [
+  isOpen: boolean,
+  dropdownRef: React.RefObject<HTMLDivElement | null>,
+  setIsOpen: React.Dispatch<React.SetStateAction<boolean>>,
+  handleItemClick: (id: number) => void
+];
+
+export const useScheduleDropdown: ScheduleDropdownType = (
+  handleSchedule: (id: number) => void
+) => {
+  const [isOpen, setIsOpen] = useState(false);
+  const dropdownRef = useRef<HTMLDivElement>(null);
+
+  const handleItemClick = (id: number) => {
+    setIsOpen(false);
+    handleSchedule(id);
+  };
+
+  useEffect(() => {
+    const handleClickOutside = (e: MouseEvent) => {
+      if (
+        dropdownRef.current &&
+        !dropdownRef.current.contains(e.target as Node)
+      ) {
+        setIsOpen(false);
+      }
+    };
+
+    if (isOpen) {
+      document.addEventListener("mousedown", handleClickOutside);
+    }
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, [isOpen]);
+
+  useEffect(() => {
+    const handleEscape = (e: KeyboardEvent) => {
+      if (e.key === "Escape") {
+        setIsOpen(false);
+      }
+    };
+
+    if (isOpen) {
+      document.addEventListener("keydown", handleEscape);
+    }
+
+    return () => {
+      document.removeEventListener("keydown", handleEscape);
+    };
+  }, [isOpen]);
+
+  return [isOpen, dropdownRef, setIsOpen, handleItemClick];
+};
