+++ src/pages/MyProfile/FileInput/custom-file-input.tsx
@@ -1,74 +1,16 @@
-import { useEffect, useRef, useState, type ChangeEvent } from "react";
-import defaultImage from "../../../assets/atc/default_profile.svg";
-import { useMyProfile } from "../../../hooks/queries";
-import { uploadProfileImage } from "../../../lib/users/api";
-import { useUpdateMyProfile } from "../../../hooks/mutations";
+import { useCustomFileInput } from "./useCustomFileInput";
 
 const CustomFileInput = () => {
-  const accessToken = localStorage.getItem("accessToken");
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const [fileImage, setFileImage] = useState<File | null>(null);
-  const [previewImage, setPreviewImage] = useState(defaultImage);
-  const { data: user, isLoading } = useMyProfile(accessToken);
-  const imageMutation = useUpdateMyProfile(accessToken!);
-
-  const handleFileClick = () => {
-    if (!fileInputRef.current) return;
-    fileInputRef.current.click();
-  };
-
-  const handleFileChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const file = e.target.files![0] ?? defaultImage;
-
-    if (!file) {
-      setFileImage(null);
-      setPreviewImage(user?.profileImageUrl ?? defaultImage);
-      return;
-    }
-
-    setFileImage(file);
-    const fileImageUrl = URL.createObjectURL(file);
-    setPreviewImage(fileImageUrl);
-  };
-
-  const handleCancel = () => {
-    if (fileInputRef.current) {
-      fileInputRef.current.value = "";
-    }
-    setFileImage(null);
-    setPreviewImage(user?.profileImageUrl ?? defaultImage);
-  };
-
-  const handleFileUpload = async () => {
-    if (fileImage) {
-      try {
-        const { profileImageUrl } = await uploadProfileImage(
-          fileImage,
-          accessToken!
-        );
-        imageMutation.mutate({ profileImageUrl });
-        setFileImage(null);
-        setPreviewImage(profileImageUrl);
-      } catch (error) {
-        console.error("프로필 이미지 저장 실패:", error);
-      }
-    }
-  };
-
-  useEffect(() => {
-    if (!isLoading && previewImage === defaultImage) {
-      setPreviewImage(user?.profileImageUrl ?? defaultImage);
-    }
-    if (
-      previewImage !== defaultImage &&
-      previewImage !== user?.profileImageUrl
-    ) {
-      return () => {
-        URL.revokeObjectURL(previewImage);
-      };
-    }
-    setPreviewImage(user?.profileImageUrl ?? defaultImage);
-  }, [user?.profileImageUrl, previewImage, isLoading]);
+  const [
+    user,
+    isLoading,
+    previewImage,
+    fileInputRef,
+    handleFileChange,
+    handleFileClick,
+    handleCancel,
+    handleFileUpload,
+  ] = useCustomFileInput();
 
   if (isLoading) {
     return (
@@ -119,7 +61,8 @@ const CustomFileInput = () => {
           onClick={handleCancel}
           className={`${
             previewImage === user?.profileImageUrl && "hidden"
-          } p-2 rounded-[8px] hover:bg-primary-100 cursor-pointer`}
+          } px-2 py-1 rounded-[8px] cursor-pointer hover:bg-primary-100
+          transition transform duration-200 hover:scale-105 hover:shadow-md`}
         >
           취소
         </button>
@@ -128,7 +71,8 @@ const CustomFileInput = () => {
           type="button"
           className={`${
             previewImage === user?.profileImageUrl && "hidden"
-          } p-2 rounded-[8px] hover:bg-primary-100 cursor-pointer`}
+          } px-2 py-1 rounded-[8px] cursor-pointer hover:bg-primary-100
+          transition transform duration-200 hover:scale-105 hover:shadow-md`}
         >
           저장
         </button>
